refactor deploy
