+++ mmyolo/deploy/__init__.py
@@ -1,2 +1,10 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from mmdeploy.version import __version__  # noqa F401
+from mmdeploy.codebase.base import BaseTask, MMCodebase, get_codebase_class
+
+from .models import *  # noqa: F401,F403
+from .object_detection import MMYOLO, YOLOObjectDetection
+
+__all__ = [
+    'MMCodebase', 'BaseTask', 'get_codebase_class', 'MMYOLO',
+    'YOLOObjectDetection'
+]
