--- mmyolo/deploy/codebase/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) OpenMMLab. All rights reserved.
-from mmdeploy.codebase.base import BaseTask, MMCodebase, get_codebase_class
-
-__all__ = ['MMCodebase', 'BaseTask', 'get_codebase_class']
