--- mmyolo/deploy/codebase/mmyolo/__init__.py
@@ -1,3 +1,0 @@
-# Copyright (c) OpenMMLab. All rights reserved.
-from .deploy import *  # noqa: F401,F403
-from .models import *  # noqa: F401,F403
