--- mmyolo/deploy/codebase/mmyolo/deploy/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) OpenMMLab. All rights reserved.
-from .object_detection import MMYOLO, YOLOObjectDetection
-
-__all__ = ['MMYOLO', 'YOLOObjectDetection']
