+++ mmyolo/deploy/models/dense_heads/yolov5_head.py
@@ -1,0 +1,1 @@
+# Copyright (c) OpenMMLab. All rights reserved.
